requirements.txt
anthropic
python-docx
flask
flask-socketio
python-dotenv
openpyxl
requests
beautifulsoup4
